Split BlittableValueType into 2 files and change namespace
