+++ src/OpenTK.OpenAL/AudioCapture.cs
@@ -27,7 +27,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
-using OpenTK.Core;
+using OpenTK.Core.Utility;
 using OpenTK.OpenAL.Native;
 
 namespace OpenTK.OpenAL
