+++ src/OpenTK.OpenAL/Native/AL/AL.cs
@@ -8,7 +8,7 @@
 using System;
 using System.Runtime.InteropServices;
 using System.Security;
-using OpenTK.Core;
+using OpenTK.Core.Utility;
 using OpenTK.Mathematics;
 
 /* Type Mapping
